+++ frb_example/pure_dart/rust/src/api/mod.rs
@@ -50,5 +50,6 @@ pub mod structure;
 pub mod tuple;
 pub mod type_alias;
 pub mod uuid_type;
+pub mod proxy;
 
 pub fn function_at_api_mod_rs() {}
